Rename '_multipartformadatencode' module to '_util'
